clean where === undefined

+++ src/modules/utils.js
@@ -3,7 +3,7 @@
  * @Author: bubao
  * @Date: 2019-09-09 16:02:22
  * @LastEditors: bubao
- * @LastEditTime: 2019-09-10 18:13:36
+ * @LastEditTime: 2019-09-10 18:20:11
  */
 
 const crypto = require("crypto");
@@ -20,10 +20,7 @@ function key() {
 }
 
 function clean(where) {
-	if (where === undefined) {
-		return;
-	}
-	if (where.indexOf("/var/tmp/") === 0) {
+	if (where !== undefined && where.indexOf("/var/tmp/") === 0) {
 		return exec(`rm -rf ${where}`);
 	}
 }
